about.html, contact.html, collection_data.html, store_location.html - Completed

+++ src/js/collection_data_script.js
@@ -1,21 +1,43 @@
-function displayAllLocalStorage() {
-    if (localStorage.getItem("enquires") !== null) {
-        var enquires;
-        var enquiryArray = localStorage.getItem("enquires").split(",");
-        var temp = [];
-        // Reassembles equires into array of arrays
-        for (var i = 0; i < enquiryArray.length; ++i) {
-            temp.push(enquiryArray[i]);
-            if (i % 5 == 4) {
-                enquires.push(temp);
-                temp = [];
+$(document).ready(function() {
+    function displayLS() {
+        if (typeof (Storage) !== "undefined") {
+            
+            // Removes display element after equiry added
+            document.getElementById("display").innerHTML = "";
+            
+            var table = "<div class=\"row\"><div class=\"head\">Customer</div><div class=\"head\">Email</div><div class=\"head\">Cell Phone Number</div><div class=\"head\">Enquiry</div><div class=\"head\">Time Stamp</div></div>";
+            
+            var enquires = [], enquiry = [];            
+            if(localStorage.getItem("enquires") !== null) {
+                var storage = localStorage.getItem("enquires").split(",");
+                // Reassembles equires into array of arrays
+                for (var i = 0; i < storage.length; ++i) {
+                    enquiry.push(storage[i]);
+                    if (i % 5 == 4) {
+                        var row = "<div class=\"row\"><div class=\"column\">" + enquiry[0] + "</div><div class=\"column\">" + enquiry[1] +"</div><div class=\"column\">" + enquiry[2] + "</div><div class=\"column\">" + enquiry[3] + "</div><div class=\"column\">" + enquiry[4] + "</div></div>"            
+                        table += row;
+                        enquiry = [];
+                    }
+                }
+                document.getElementById("table").innerHTML = table;
             }
-        }        
-    } else {
-        document.getElementById("display").innerHTML = "There are no customer equires";
+        }
     }
-}
+    $("#disLSBtn").click(displayLS);
+});
 
-function clearAllLocalStorage() {
-    localStorage.clear();
-}
+//
+//    function test() {
+//        document.getElementById("table").innerHTML += "<div class=\"row\"><div class=\"column\">x</div><div class=\"column\">x</div><div class=\"column\">x</div><div class=\"column\">x</div><div class=\"column\">x</div></div>";
+//    }
+//    $("#clearLSBtn").click(test);
+
+
+$(document).ready(function() {
+    function clearLS() {
+        if (typeof (Storage) !== "undefined") {
+            localStorage.clear();            
+        }
+    }
+    $("#clearLSBtn").click(clearLS);
+});
